Restore server.js
README update
License cleanup
More logs in bot.js
Fast fix for css file
