+++ README.md
@@ -6,4 +6,6 @@ Uruchamiany na platfomie Glitch.com.
 W celu uruchomienia bota (podczas uśpienia po nieaktywności) należy skorzystać z linku
 https://kacprzykbot.glitch.me
 
-Użyto Glitch Github Sync - https://github.com/nmcardoso/glitch-github-sync
+Użyto 
+- [Glitch Github Sync](https://github.com/nmcardoso/glitch-github-sync)
+- [W3.CSS](https://www.w3schools.com/w3css/default.asp)
