+++ bot.js
@@ -4,6 +4,7 @@
 ///  Bot nie ma na celu obrażania Pana Cezarego
 ///  Jest używany tylko na serwerze Testowym i prywatnej społeczności w celach humorystycznych i edukacyjnych
 ///
+console.log("Trwa inicjalizacja...");
 const Discord = require("discord.js");
 const client = new Discord.Client();
 const lubisz = [
@@ -32,19 +33,15 @@ const ct = [
   "Ahh te czasy...",
   "Może lepiej zacznijmy lekcje..."
 ];
-var slowo;
-client.login(process.env.TOKEN);
-console.log("Włączam");
 
 function los(slowa) {
   return slowa[Math.floor(Math.random() * slowa.length)];
 }
 
-client.on("ready", readyDiscord);
-function readyDiscord() {
+client.on("ready", () => {
   console.log("o działa");
   client.user.setActivity("okolice za oknem", { type: "WATCHING" });
-}
+});
 
 client.on("message", msg => {
   if (msg.author != "810790057459187733") {
@@ -150,46 +147,7 @@ client.on("message", msg => {
   }
 });
 
-const express = require("express");
-const cmd = require("node-cmd");
-const crypto = require("crypto");
-const bodyParser = require("body-parser");
-
-const app = express();
-app.use(bodyParser.json());
-
-const verifySignature = (req, res, next) => {
-  const payload = JSON.stringify(req.body);
-  const hmac = crypto.createHmac("sha1", process.env.GITHUB_SECRET);
-  const digest = "sha1=" + hmac.update(payload).digest("hex");
-  const checksum = req.headers["x-hub-signature"];
-
-  if (!checksum || !digest || checksum !== digest) {
-    return res.status(403).send("auth failed");
-  }
-
-  return next();
-};
-
-// Github webhook listener
-app.post("/git", verifySignature, (req, res) => {
-  if (req.headers["x-github-event"] == "push") {
-    cmd.get("bash git.sh", (err, data) => {
-      if (err) return console.log(err);
-      console.log(data);
-      return res.status(200).send(data);
-    });
-  } else if (req.headers["x-github-event"] == "ping") {
-    return res.status(200).send("PONG");
-  } else {
-    return res.status(200).send("Unsuported Github event. Nothing done.");
-  }
-});
-
-app.get("/", (req, res) => {
-  res.sendFile(__dirname + "/views/index.html");
-});
-
-app.listen(process.env.PORT, () => {
-  console.log(`Your app is listening on port ${process.env.PORT}`);
-});
+console.log("Logowanie się jako bot...");
+client.login(process.env.TOKEN);
+console.log("Uruchamianie serwera www...");
+require('./server.js');
